+++ source/common/upstream/load_balancer_impl.cc
@@ -784,16 +784,25 @@ SubsetSelectorImpl::SubsetSelectorImpl(
       fallback_keys_subset_(fallback_keys_subset.begin(), fallback_keys_subset.end()) {
 
   if (fallback_policy_ != envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::KEYS_SUBSET) {
+    // defining fallback_keys_subset_ for a fallback policy other than KEYS_SUBSET doesn't have
+    // any effect and it is probably a user mistake. We should let the user know about it.
     if (!fallback_keys_subset_.empty()) {
       throw EnvoyException("fallback_keys_subset can be set only for KEYS_SUBSET fallback_policy");
     }
     return;
   }
 
+  // if KEYS_SUBSET fallback policy is selected, fallback_keys_subset must not be empty, because
+  // it would be the same as not defining fallback policy at all (global fallback policy would be
+  // used)
   if (fallback_keys_subset_.empty()) {
     throw EnvoyException("fallback_keys_subset cannot be empty");
   }
 
+  // We allow only for a fallback to a subset of the selector keys because this is probably the
+  // only use case that makes sense (fallback from more specific selector to less specific
+  // selector). Potentially we can relax this constraint in the future if there will be a use case
+  // for this.
   if (!std::includes(selector_keys_.begin(), selector_keys_.end(), fallback_keys_subset_.begin(),
                      fallback_keys_subset_.end())) {
     throw EnvoyException("fallback_keys_subset must be a subset of selector keys");