+++ test/common/upstream/subset_lb_test.cc
@@ -98,31 +98,6 @@ private:
   std::vector<Router::MetadataMatchCriterionConstSharedPtr> matches_;
 };
 
-class TestSubsetSelector : public SubsetSelector {
-public:
-  TestSubsetSelector(
-      const std::set<std::string>& selector_keys,
-      envoy::api::v2::Cluster_LbSubsetConfig_LbSubsetSelector::LbSubsetSelectorFallbackPolicy
-          fallback_policy,
-      const std::set<std::string>& fallback_keys_subset)
-      : selector_keys_(selector_keys), fallback_policy_(fallback_policy),
-        fallback_keys_subset_(fallback_keys_subset) {}
-
-  // SubsetSelector
-  const std::set<std::string>& selectorKeys() const override { return selector_keys_; }
-  envoy::api::v2::Cluster_LbSubsetConfig_LbSubsetSelector::LbSubsetSelectorFallbackPolicy
-  fallbackPolicy() const override {
-    return fallback_policy_;
-  }
-  const std::set<std::string>& fallbackKeysSubset() const override { return fallback_keys_subset_; }
-
-private:
-  const std::set<std::string> selector_keys_;
-  const envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::LbSubsetSelectorFallbackPolicy
-      fallback_policy_;
-  const std::set<std::string> fallback_keys_subset_;
-};
-
 class TestLoadBalancerContext : public LoadBalancerContextBase {
 public:
   TestLoadBalancerContext(
@@ -298,6 +273,33 @@ public:
     }
 
     return default_subset;
+  }
+
+  SubsetSelectorPtr makeSelector(
+      const std::set<std::string>& selector_keys,
+      envoy::api::v2::Cluster_LbSubsetConfig_LbSubsetSelector::LbSubsetSelectorFallbackPolicy
+          fallback_policy,
+      const std::set<std::string>& fallback_keys_subset) {
+
+    Protobuf::RepeatedPtrField<std::string> selector_keys_mapped;
+    for (const auto& it : selector_keys) {
+      selector_keys_mapped.Add(std::string(it));
+    }
+
+    Protobuf::RepeatedPtrField<std::string> fallback_keys_subset_mapped;
+    for (const auto& it : fallback_keys_subset) {
+      fallback_keys_subset_mapped.Add(std::string(it));
+    }
+
+    return std::make_shared<SubsetSelectorImpl>(selector_keys_mapped, fallback_policy,
+                                                fallback_keys_subset_mapped);
+  }
+
+  SubsetSelectorPtr makeSelector(
+      const std::set<std::string>& selector_keys,
+      envoy::api::v2::Cluster_LbSubsetConfig_LbSubsetSelector::LbSubsetSelectorFallbackPolicy
+          fallback_policy) {
+    return makeSelector(selector_keys, fallback_policy, {});
   }
 
   void modifyHosts(HostVector add, HostVector remove, absl::optional<uint32_t> add_in_locality = {},
@@ -619,10 +621,8 @@ TEST_F(SubsetLoadBalancerTest, BalancesSubset) {
   EXPECT_CALL(subset_info_, fallbackPolicy())
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::NO_FALLBACK));
 
-  std::vector<SubsetSelectorPtr> subset_selectors = {std::make_shared<TestSubsetSelector>(
-      TestSubsetSelector{{"version"},
-                         envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                         {}})};
+  std::vector<SubsetSelectorPtr> subset_selectors = {makeSelector(
+      {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
 
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
@@ -648,10 +648,8 @@ TEST_P(SubsetLoadBalancerTest, BalancesSubsetAfterUpdate) {
   EXPECT_CALL(subset_info_, fallbackPolicy())
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::NO_FALLBACK));
 
-  std::vector<SubsetSelectorPtr> subset_selectors = {std::make_shared<TestSubsetSelector>(
-      TestSubsetSelector{{"version"},
-                         envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                         {}})};
+  std::vector<SubsetSelectorPtr> subset_selectors = {makeSelector(
+      {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
 
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
@@ -689,10 +687,8 @@ TEST_P(SubsetLoadBalancerTest, ListAsAnyEnabled) {
   EXPECT_CALL(subset_info_, fallbackPolicy())
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::NO_FALLBACK));
 
-  std::vector<SubsetSelectorPtr> subset_selectors = {std::make_shared<TestSubsetSelector>(
-      TestSubsetSelector{{"version"},
-                         envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                         {}})};
+  std::vector<SubsetSelectorPtr> subset_selectors = {makeSelector(
+      {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
   EXPECT_CALL(subset_info_, listAsAny()).WillRepeatedly(Return(true));
 
@@ -718,10 +714,8 @@ TEST_P(SubsetLoadBalancerTest, ListAsAnyEnabledMultipleLists) {
   EXPECT_CALL(subset_info_, fallbackPolicy())
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::NO_FALLBACK));
 
-  std::vector<SubsetSelectorPtr> subset_selectors = {std::make_shared<TestSubsetSelector>(
-      TestSubsetSelector{{"version"},
-                         envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                         {}})};
+  std::vector<SubsetSelectorPtr> subset_selectors = {makeSelector(
+      {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
   EXPECT_CALL(subset_info_, listAsAny()).WillRepeatedly(Return(true));
 
@@ -759,10 +753,9 @@ TEST_P(SubsetLoadBalancerTest, ListAsAnyEnabledMultipleListsForSingleHost) {
   EXPECT_CALL(subset_info_, fallbackPolicy())
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::NO_FALLBACK));
 
-  std::vector<SubsetSelectorPtr> subset_selectors = {std::make_shared<TestSubsetSelector>(
-      TestSubsetSelector{{"version", "hardware"},
-                         envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                         {}})};
+  std::vector<SubsetSelectorPtr> subset_selectors = {
+      makeSelector({"version", "hardware"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
   EXPECT_CALL(subset_info_, listAsAny()).WillRepeatedly(Return(true));
 
@@ -800,10 +793,8 @@ TEST_P(SubsetLoadBalancerTest, ListAsAnyDisable) {
   EXPECT_CALL(subset_info_, fallbackPolicy())
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::NO_FALLBACK));
 
-  std::vector<SubsetSelectorPtr> subset_selectors = {std::make_shared<TestSubsetSelector>(
-      TestSubsetSelector{{"version"},
-                         envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                         {}})};
+  std::vector<SubsetSelectorPtr> subset_selectors = {makeSelector(
+      {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
   init({});
@@ -825,10 +816,8 @@ TEST_P(SubsetLoadBalancerTest, UpdateFailover) {
   EXPECT_CALL(subset_info_, fallbackPolicy())
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::NO_FALLBACK));
 
-  std::vector<SubsetSelectorPtr> subset_selectors = {std::make_shared<TestSubsetSelector>(
-      TestSubsetSelector{{"version"},
-                         envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                         {}})};
+  std::vector<SubsetSelectorPtr> subset_selectors = {makeSelector(
+      {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
 
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
@@ -859,12 +848,10 @@ TEST_P(SubsetLoadBalancerTest, OnlyMetadataChanged) {
   TestLoadBalancerContext context_default({{"default", "true"}});
 
   std::vector<SubsetSelectorPtr> subset_selectors = {
-      std::make_shared<TestSubsetSelector>(TestSubsetSelector{
-          {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED, {}}),
-      std::make_shared<TestSubsetSelector>(
-          TestSubsetSelector{{"default"},
-                             envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                             {}})};
+      makeSelector({"version"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED),
+      makeSelector({"default"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
 
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
@@ -954,12 +941,10 @@ TEST_P(SubsetLoadBalancerTest, MetadataChangedHostsAddedRemoved) {
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::DEFAULT_SUBSET));
 
   std::vector<SubsetSelectorPtr> subset_selectors = {
-      std::make_shared<TestSubsetSelector>(TestSubsetSelector{
-          {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED, {}}),
-      std::make_shared<TestSubsetSelector>(
-          TestSubsetSelector{{"default"},
-                             envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                             {}})};
+      makeSelector({"version"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED),
+      makeSelector({"default"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
   // Add hosts initial hosts.
@@ -1038,10 +1023,8 @@ TEST_P(SubsetLoadBalancerTest, UpdateRemovingLastSubsetHost) {
   EXPECT_CALL(subset_info_, fallbackPolicy())
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::ANY_ENDPOINT));
 
-  std::vector<SubsetSelectorPtr> subset_selectors = {std::make_shared<TestSubsetSelector>(
-      TestSubsetSelector{{"version"},
-                         envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                         {}})};
+  std::vector<SubsetSelectorPtr> subset_selectors = {makeSelector(
+      {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
   init({
@@ -1075,14 +1058,10 @@ TEST_P(SubsetLoadBalancerTest, UpdateRemovingUnknownHost) {
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::NO_FALLBACK));
 
   std::vector<SubsetSelectorPtr> subset_selectors = {
-      std::make_shared<TestSubsetSelector>(
-          TestSubsetSelector{{"stage", "version"},
-                             envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                             {}}),
-      std::make_shared<TestSubsetSelector>(
-          TestSubsetSelector{{"version"},
-                             envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                             {}})};
+      makeSelector({"stage", "version"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED),
+      makeSelector({"version"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
 
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
@@ -1106,12 +1085,10 @@ TEST_F(SubsetLoadBalancerTest, UpdateModifyingOnlyHostHealth) {
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::NO_FALLBACK));
 
   std::vector<SubsetSelectorPtr> subset_selectors = {
-      std::make_shared<TestSubsetSelector>(TestSubsetSelector{
-          {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED, {}}),
-      std::make_shared<TestSubsetSelector>(
-          TestSubsetSelector{{"hardware"},
-                             envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                             {}})};
+      makeSelector({"version"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED),
+      makeSelector({"hardware"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
 
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
@@ -1148,12 +1125,10 @@ TEST_F(SubsetLoadBalancerTest, BalancesDisjointSubsets) {
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::NO_FALLBACK));
 
   std::vector<SubsetSelectorPtr> subset_selectors = {
-      std::make_shared<TestSubsetSelector>(TestSubsetSelector{
-          {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED, {}}),
-      std::make_shared<TestSubsetSelector>(
-          TestSubsetSelector{{"hardware"},
-                             envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                             {}})};
+      makeSelector({"version"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED),
+      makeSelector({"hardware"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
 
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
@@ -1178,14 +1153,10 @@ TEST_F(SubsetLoadBalancerTest, BalancesOverlappingSubsets) {
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::NO_FALLBACK));
 
   std::vector<SubsetSelectorPtr> subset_selectors = {
-      std::make_shared<TestSubsetSelector>(
-          TestSubsetSelector{{"stage", "version"},
-                             envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                             {}}),
-      std::make_shared<TestSubsetSelector>(
-          TestSubsetSelector{{"version"},
-                             envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                             {}})};
+      makeSelector({"stage", "version"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED),
+      makeSelector({"version"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
 
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
@@ -1221,12 +1192,10 @@ TEST_F(SubsetLoadBalancerTest, BalancesNestedSubsets) {
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::NO_FALLBACK));
 
   std::vector<SubsetSelectorPtr> subset_selectors = {
-      std::make_shared<TestSubsetSelector>(
-          TestSubsetSelector{{"stage", "version"},
-                             envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                             {}}),
-      std::make_shared<TestSubsetSelector>(TestSubsetSelector{
-          {"stage"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED, {}})};
+      makeSelector({"stage", "version"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED),
+      makeSelector({"stage"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
 
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
@@ -1259,10 +1228,8 @@ TEST_F(SubsetLoadBalancerTest, IgnoresUnselectedMetadata) {
   EXPECT_CALL(subset_info_, fallbackPolicy())
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::NO_FALLBACK));
 
-  std::vector<SubsetSelectorPtr> subset_selectors = {std::make_shared<TestSubsetSelector>(
-      TestSubsetSelector{{"version"},
-                         envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                         {}})};
+  std::vector<SubsetSelectorPtr> subset_selectors = {makeSelector(
+      {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
 
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
@@ -1286,10 +1253,8 @@ TEST_F(SubsetLoadBalancerTest, IgnoresHostsWithoutMetadata) {
   EXPECT_CALL(subset_info_, fallbackPolicy())
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::NO_FALLBACK));
 
-  std::vector<SubsetSelectorPtr> subset_selectors = {std::make_shared<TestSubsetSelector>(
-      TestSubsetSelector{{"version"},
-                         envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                         {}})};
+  std::vector<SubsetSelectorPtr> subset_selectors = {makeSelector(
+      {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
 
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
@@ -1337,8 +1302,7 @@ TEST_F(SubsetLoadBalancerTest, ZoneAwareFallback) {
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::ANY_ENDPOINT));
 
   std::vector<SubsetSelectorPtr> subset_selectors = {
-      std::make_shared<TestSubsetSelector>(TestSubsetSelector{
-          {"x"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED, {}})};
+      makeSelector({"x"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
 
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
@@ -1384,8 +1348,7 @@ TEST_P(SubsetLoadBalancerTest, ZoneAwareFallbackAfterUpdate) {
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::ANY_ENDPOINT));
 
   std::vector<SubsetSelectorPtr> subset_selectors = {
-      std::make_shared<TestSubsetSelector>(TestSubsetSelector{
-          {"x"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED, {}})};
+      makeSelector({"x"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
 
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
@@ -1445,10 +1408,8 @@ TEST_F(SubsetLoadBalancerTest, ZoneAwareFallbackDefaultSubset) {
   const ProtobufWkt::Struct default_subset = makeDefaultSubset({{"version", "default"}});
   EXPECT_CALL(subset_info_, defaultSubset()).WillRepeatedly(ReturnRef(default_subset));
 
-  std::vector<SubsetSelectorPtr> subset_selectors = {std::make_shared<TestSubsetSelector>(
-      TestSubsetSelector{{"version"},
-                         envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                         {}})};
+  std::vector<SubsetSelectorPtr> subset_selectors = {makeSelector(
+      {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
 
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
@@ -1503,10 +1464,8 @@ TEST_P(SubsetLoadBalancerTest, ZoneAwareFallbackDefaultSubsetAfterUpdate) {
   const ProtobufWkt::Struct default_subset = makeDefaultSubset({{"version", "default"}});
   EXPECT_CALL(subset_info_, defaultSubset()).WillRepeatedly(ReturnRef(default_subset));
 
-  std::vector<SubsetSelectorPtr> subset_selectors = {std::make_shared<TestSubsetSelector>(
-      TestSubsetSelector{{"version"},
-                         envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                         {}})};
+  std::vector<SubsetSelectorPtr> subset_selectors = {makeSelector(
+      {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
 
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
@@ -1571,10 +1530,8 @@ TEST_F(SubsetLoadBalancerTest, ZoneAwareBalancesSubsets) {
   EXPECT_CALL(subset_info_, fallbackPolicy())
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::NO_FALLBACK));
 
-  std::vector<SubsetSelectorPtr> subset_selectors = {std::make_shared<TestSubsetSelector>(
-      TestSubsetSelector{{"version"},
-                         envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                         {}})};
+  std::vector<SubsetSelectorPtr> subset_selectors = {makeSelector(
+      {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
   EXPECT_CALL(runtime_.snapshot_, getInteger("upstream.healthy_panic_threshold", 50))
@@ -1627,10 +1584,8 @@ TEST_P(SubsetLoadBalancerTest, ZoneAwareBalancesSubsetsAfterUpdate) {
   EXPECT_CALL(subset_info_, fallbackPolicy())
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::NO_FALLBACK));
 
-  std::vector<SubsetSelectorPtr> subset_selectors = {std::make_shared<TestSubsetSelector>(
-      TestSubsetSelector{{"version"},
-                         envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                         {}})};
+  std::vector<SubsetSelectorPtr> subset_selectors = {makeSelector(
+      {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
   EXPECT_CALL(runtime_.snapshot_, getInteger("upstream.healthy_panic_threshold", 50))
@@ -1787,10 +1742,8 @@ TEST_F(SubsetLoadBalancerTest, EnabledLocalityWeightAwareness) {
 
 TEST_F(SubsetLoadBalancerTest, EnabledScaleLocalityWeights) {
 
-  std::vector<SubsetSelectorPtr> subset_selectors = {std::make_shared<TestSubsetSelector>(
-      TestSubsetSelector{{"version"},
-                         envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                         {}})};
+  std::vector<SubsetSelectorPtr> subset_selectors = {makeSelector(
+      {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
   EXPECT_CALL(subset_info_, isEnabled()).WillRepeatedly(Return(true));
   EXPECT_CALL(subset_info_, localityWeightAware()).WillRepeatedly(Return(true));
@@ -1832,10 +1785,8 @@ TEST_F(SubsetLoadBalancerTest, EnabledScaleLocalityWeights) {
 
 TEST_F(SubsetLoadBalancerTest, EnabledScaleLocalityWeightsRounding) {
 
-  std::vector<SubsetSelectorPtr> subset_selectors = {std::make_shared<TestSubsetSelector>(
-      TestSubsetSelector{{"version"},
-                         envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                         {}})};
+  std::vector<SubsetSelectorPtr> subset_selectors = {makeSelector(
+      {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
   EXPECT_CALL(subset_info_, isEnabled()).WillRepeatedly(Return(true));
   EXPECT_CALL(subset_info_, localityWeightAware()).WillRepeatedly(Return(true));
@@ -1873,10 +1824,8 @@ TEST_F(SubsetLoadBalancerTest, EnabledScaleLocalityWeightsRounding) {
 
 // Regression for bug where missing locality weights crashed scaling and locality aware subset LBs.
 TEST_F(SubsetLoadBalancerTest, ScaleLocalityWeightsWithNoLocalityWeights) {
-  std::vector<SubsetSelectorPtr> subset_selectors = {std::make_shared<TestSubsetSelector>(
-      TestSubsetSelector{{"version"},
-                         envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                         {}})};
+  std::vector<SubsetSelectorPtr> subset_selectors = {makeSelector(
+      {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
   EXPECT_CALL(subset_info_, isEnabled()).WillRepeatedly(Return(true));
   EXPECT_CALL(subset_info_, localityWeightAware()).WillRepeatedly(Return(true));
@@ -1898,10 +1847,8 @@ TEST_P(SubsetLoadBalancerTest, GaugesUpdatedOnDestroy) {
   EXPECT_CALL(subset_info_, fallbackPolicy())
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::ANY_ENDPOINT));
 
-  std::vector<SubsetSelectorPtr> subset_selectors = {std::make_shared<TestSubsetSelector>(
-      TestSubsetSelector{{"version"},
-                         envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED,
-                         {}})};
+  std::vector<SubsetSelectorPtr> subset_selectors = {makeSelector(
+      {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
   init({
@@ -1921,10 +1868,8 @@ TEST_P(SubsetLoadBalancerTest, SubsetSelectorNoFallbackPerSelector) {
   EXPECT_CALL(subset_info_, fallbackPolicy())
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::DEFAULT_SUBSET));
 
-  std::vector<SubsetSelectorPtr> subset_selectors = {std::make_shared<TestSubsetSelector>(
-      TestSubsetSelector{{"version"},
-                         envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NO_FALLBACK,
-                         {}})};
+  std::vector<SubsetSelectorPtr> subset_selectors = {makeSelector(
+      {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NO_FALLBACK)};
 
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
@@ -1953,12 +1898,10 @@ TEST_P(SubsetLoadBalancerTest, FallbackNotDefinedForIntermediateSelector) {
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::ANY_ENDPOINT));
 
   std::vector<SubsetSelectorPtr> subset_selectors = {
-      std::make_shared<TestSubsetSelector>(TestSubsetSelector{
-          {"stage"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED, {}}),
-      std::make_shared<TestSubsetSelector>(TestSubsetSelector{
-          {"stage", "version"},
-          envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::ANY_ENDPOINT,
-          {}})};
+      makeSelector({"stage"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED),
+      makeSelector({"stage", "version"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::ANY_ENDPOINT)};
 
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
@@ -1978,10 +1921,8 @@ TEST_P(SubsetLoadBalancerTest, SubsetSelectorFallbackOverridesTopLevelOne) {
   EXPECT_CALL(subset_info_, fallbackPolicy())
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::ANY_ENDPOINT));
 
-  std::vector<SubsetSelectorPtr> subset_selectors = {std::make_shared<TestSubsetSelector>(
-      TestSubsetSelector{{"version"},
-                         envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NO_FALLBACK,
-                         {}})};
+  std::vector<SubsetSelectorPtr> subset_selectors = {makeSelector(
+      {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NO_FALLBACK)};
 
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
@@ -1998,10 +1939,8 @@ TEST_P(SubsetLoadBalancerTest, SubsetSelectorNoFallbackMatchesTopLevelOne) {
   EXPECT_CALL(subset_info_, fallbackPolicy())
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::NO_FALLBACK));
 
-  std::vector<SubsetSelectorPtr> subset_selectors = {std::make_shared<TestSubsetSelector>(
-      TestSubsetSelector{{"version"},
-                         envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NO_FALLBACK,
-                         {}})};
+  std::vector<SubsetSelectorPtr> subset_selectors = {makeSelector(
+      {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NO_FALLBACK)};
 
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
@@ -2020,14 +1959,12 @@ TEST_P(SubsetLoadBalancerTest, SubsetSelectorDefaultAnyFallbackPerSelector) {
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::NO_FALLBACK));
 
   std::vector<SubsetSelectorPtr> subset_selectors = {
-      std::make_shared<TestSubsetSelector>(TestSubsetSelector{
-          {"version"},
-          envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::DEFAULT_SUBSET,
-          {}}),
-      std::make_shared<TestSubsetSelector>(TestSubsetSelector{
-          {"app"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::ANY_ENDPOINT, {}}),
-      std::make_shared<TestSubsetSelector>(TestSubsetSelector{
-          {"foo"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED, {}})};
+      makeSelector({"version"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::DEFAULT_SUBSET),
+      makeSelector({"app"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::ANY_ENDPOINT),
+      makeSelector({"foo"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED)};
 
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
@@ -2060,10 +1997,8 @@ TEST_P(SubsetLoadBalancerTest, SubsetSelectorDefaultAfterUpdate) {
   const ProtobufWkt::Struct default_subset = makeDefaultSubset({{"version", "default"}});
   EXPECT_CALL(subset_info_, defaultSubset()).WillRepeatedly(ReturnRef(default_subset));
 
-  std::vector<SubsetSelectorPtr> subset_selectors = {std::make_shared<TestSubsetSelector>(
-      TestSubsetSelector{{"version"},
-                         envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::DEFAULT_SUBSET,
-                         {}})};
+  std::vector<SubsetSelectorPtr> subset_selectors = {makeSelector(
+      {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::DEFAULT_SUBSET)};
 
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
@@ -2088,10 +2023,8 @@ TEST_P(SubsetLoadBalancerTest, SubsetSelectorAnyAfterUpdate) {
   EXPECT_CALL(subset_info_, fallbackPolicy())
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::NO_FALLBACK));
 
-  std::vector<SubsetSelectorPtr> subset_selectors = {std::make_shared<TestSubsetSelector>(
-      TestSubsetSelector{{"version"},
-                         envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::ANY_ENDPOINT,
-                         {}})};
+  std::vector<SubsetSelectorPtr> subset_selectors = {makeSelector(
+      {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::ANY_ENDPOINT)};
 
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
@@ -2118,20 +2051,15 @@ TEST_P(SubsetLoadBalancerTest, FallbackForCompoundSelector) {
   EXPECT_CALL(subset_info_, defaultSubset()).WillRepeatedly(ReturnRef(default_subset));
 
   std::vector<SubsetSelectorPtr> subset_selectors = {
-      std::make_shared<TestSubsetSelector>(TestSubsetSelector{
-          {"version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED, {}}),
-      std::make_shared<TestSubsetSelector>(
-          TestSubsetSelector{{"version", "hardware", "stage"},
-                             envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NO_FALLBACK,
-                             {}}),
-      std::make_shared<TestSubsetSelector>(TestSubsetSelector{
-          {"version", "hardware"},
-          envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::DEFAULT_SUBSET,
-          {}}),
-      std::make_shared<TestSubsetSelector>(
-          TestSubsetSelector{{"version", "stage"},
-                             envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::KEYS_SUBSET,
-                             {"version"}})};
+      makeSelector({"version"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NOT_DEFINED),
+      makeSelector({"version", "hardware", "stage"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NO_FALLBACK),
+      makeSelector({"version", "hardware"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::DEFAULT_SUBSET),
+      makeSelector({"version", "stage"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::KEYS_SUBSET,
+                   {"version"})};
 
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
@@ -2171,16 +2099,14 @@ TEST_P(SubsetLoadBalancerTest, KeysSubsetFallbackChained) {
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::NO_FALLBACK));
 
   std::vector<SubsetSelectorPtr> subset_selectors = {
-      std::make_shared<TestSubsetSelector>(TestSubsetSelector{
-          {"stage"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NO_FALLBACK, {}}),
-      std::make_shared<TestSubsetSelector>(
-          TestSubsetSelector{{"stage", "version"},
-                             envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::KEYS_SUBSET,
-                             {"stage"}}),
-      std::make_shared<TestSubsetSelector>(
-          TestSubsetSelector{{"stage", "version", "hardware"},
-                             envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::KEYS_SUBSET,
-                             {"version", "stage"}})};
+      makeSelector({"stage"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::NO_FALLBACK),
+      makeSelector({"stage", "version"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::KEYS_SUBSET,
+                   {"stage"}),
+      makeSelector({"stage", "version", "hardware"},
+                   envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::KEYS_SUBSET,
+                   {"version", "stage"})};
 
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));
 
@@ -2207,10 +2133,9 @@ TEST_P(SubsetLoadBalancerTest, KeysSubsetFallbackToNotExistingSelector) {
   EXPECT_CALL(subset_info_, fallbackPolicy())
       .WillRepeatedly(Return(envoy::api::v2::Cluster::LbSubsetConfig::ANY_ENDPOINT));
 
-  std::vector<SubsetSelectorPtr> subset_selectors = {std::make_shared<TestSubsetSelector>(
-      TestSubsetSelector{{"stage", "version"},
-                         envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::KEYS_SUBSET,
-                         {"stage"}})};
+  std::vector<SubsetSelectorPtr> subset_selectors = {makeSelector(
+      {"stage", "version"}, envoy::api::v2::Cluster::LbSubsetConfig::LbSubsetSelector::KEYS_SUBSET,
+      {"stage"})};
 
   EXPECT_CALL(subset_info_, subsetSelectors()).WillRepeatedly(ReturnRef(subset_selectors));